+++ PracticaFinal.IPC1/src/practicafinal/ipc1/DisparoEnemIzquierda.java
@@ -21,7 +21,7 @@ public class DisparoEnemIzquierda extends TimerTask {
     private NombreJugador lista;
     private Nodo2<NombreAuto> elemento2;
     private NombreAuto nombre;
-    private int valor;
+    private double valor;
     
     
     @Override
@@ -38,14 +38,14 @@ public class DisparoEnemIzquierda extends TimerTask {
                     mapa[i][j].setIcon(autoTanque);
                     JOptionPane.showMessageDialog(null, "Un enemigo ha acertado un golpe contra ti :/");
                     valor = valor - 10;
-                    vidaAuto.setText(Integer.toString(valor));
+                    vidaAuto.setText(Double.toString(valor));
                     nombre.setVida(valor);
                     this.cancel();
                 } else if (valores[i][j]==2){
                     mapa[i][j].setIcon(autoAvion);
                     JOptionPane.showMessageDialog(null, "Un enemigo ha acertado un golpe contra ti :/");
                     valor = valor - 10;
-                    vidaAuto.setText(Integer.toString(valor));
+                    vidaAuto.setText(Double.toString(valor));
                     nombre.setVida(valor);
                     this.cancel();
                 }
